pyproject.toml
[project]
name = "generative-ai-api"
version = "0.1.0"
description = "Generative AI API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = []

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
